requirements.txt
streamlit~=1.32.2
numpy~=1.26.4
requests~=2.31.0
beautifulsoup4~=4.12.3
python-whois~=0.9.3
scikit-learn~=1.4.1.post1